Moves CombinedFilteredAggregationContext to be static nested class of AggregationFunctionUtils
